requirements.txt
python-dotenv
langchain_groq
langchain_community
langgraph
streamlit
ipykernel
tavily-python